YKW4 Templates
Itemconfig now uses the SkillID enum instead of the AbilityID enum.
BattlePassive is for parsing battle_passive_skill.cfg_7.00.18.bin. Due to how the file is, the template is currently a little messy.
BattleTechnic is for parsing battle_technic_0.13.62.cfg.bin.
The AbilityID enum was removed in favor for the SkillID enum due to files. The MoveID enum was fixed. Unused moves and Skills that were previously commented out in case they were found have been removed, since they go entirely unused in the game.